requirements.txt
pandas
selenium
bs4
numpy
matplotlib
seaborn
requests
folium
langdetect
nltk
textblob_de
openpyxl
geopandas
psycopg2-binary
sqlalchemy